long id changed to int

+++ UniAtHome/UniAtHome.DAL/Entities/Lesson.cs
@@ -4,7 +4,7 @@ namespace UniAtHome.DAL.Entities
 {
     public class Lesson : BaseAuditableEntity
     {
-        public long CourseId { get; set; }
+        public int CourseId { get; set; }
 
         public Course Course { get; set; }
     }
